+++ src/context.ts
@@ -4,18 +4,18 @@ let synthesisServerUrl: string | undefined = ''
 
 const synthesisServerHeaders: Record<string, string> = {
   'Content-Type': 'application/json',
-  'x-access-token': `${apiKey}`
+  'x-access-token': `${apiKey}`,
 }
 
 const headers: Record<string, string> = {
   'Content-Type': 'application/json',
-  'Authorization': `Token token=${apiKey}`
+  Authorization: `Token token=${apiKey}`,
 }
 
 export const context = {
   headers: () => headers,
   synthesisServerHeaders: () => synthesisServerHeaders,
-  
+
   setBaseUrl: (url: string) => {
     baseUrl = url
 
@@ -29,22 +29,22 @@ export const context = {
     headers['Authorization'] = `Token token=${key}`
     synthesisServerHeaders['x-access-token'] = key
   },
-  
+
   setSynthesisUrl: (url: string) => {
     synthesisServerUrl = url
-    
+
     if (!url.endsWith('/')) {
       synthesisServerUrl += '/'
     }
   },
-  
+
   endpoint: (version: string, endpoint: string): string => {
     let ending = endpoint.startsWith('/') ? endpoint.substring(1) : endpoint
     return `${baseUrl}${version}/${ending}`
   },
-  
+
   synServerUrl: (endpoint) => {
     let ending = endpoint.startsWith('/') ? endpoint.substring(1) : endpoint
     return `${synthesisServerUrl}${ending}`
-  }
-}  
+  },
+}