+++ src/v2/clips.ts
@@ -1,5 +1,10 @@
-import UtilV2, { ErrorResponseV2, PaginationResponseV2, ReadResponseV2, WriteResponseV2 } from "./util"
-// import { DEFAULT_BUFFER_SIZE, StreamDecoder } from './StreamDecoder'
+import UtilV2, {
+  ErrorResponseV2,
+  PaginationResponseV2,
+  ReadResponseV2,
+  WriteResponseV2,
+} from './util'
+import { DEFAULT_BUFFER_SIZE, StreamDecoder } from './StreamDecoder'
 
 export interface Clip {
   uuid: string
@@ -14,7 +19,6 @@ export interface Clip {
   created_at: Date
   updated_at: Date
 }
-
 
 interface ClipInput {
   title?: string
@@ -46,28 +50,28 @@ export interface DirectClipInput {
 }
 
 export interface DirectClip {
-  success: true;
-  audio_content: string;
+  success: true
+  audio_content: string
   audio_timestamps: {
-    graph_chars: string[];
-    graph_times: [number, number][];
-    phon_chars: string[];
-    phon_times: [number, number][];
-  };
-  duration: number;
-  synth_duration: number;
-  output_format: 'wav' | 'mp3';
-  sample_rate: number;
-  issues: string[];
+    graph_chars: string[]
+    graph_times: [number, number][]
+    phon_chars: string[]
+    phon_times: [number, number][]
+  }
+  duration: number
+  synth_duration: number
+  output_format: 'wav' | 'mp3'
+  sample_rate: number
+  issues: string[]
 }
 
 export interface DirectClipError {
-  success: false;
-  issues?: string[];
-  error_name: string;
-  error_params: unknown;
-  feedback_uuid: string;
-  message: string;
+  success: false
+  issues?: string[]
+  error_name: string
+  error_params: unknown
+  feedback_uuid: string
+  message: string
 }
 
 export interface StreamInput {
@@ -78,7 +82,10 @@ export interface StreamInput {
 
 const create = async (projectUuid: string, clipInput: ClipInput) => {
   try {
-    const response = await UtilV2.post(`projects/${projectUuid}/clips`, clipInput)
+    const response = await UtilV2.post(
+      `projects/${projectUuid}/clips`,
+      clipInput,
+    )
     let json = await response.json()
     if (json.success) {
       json = {
@@ -86,8 +93,8 @@ const create = async (projectUuid: string, clipInput: ClipInput) => {
         item: {
           ...json.item,
           created_at: new Date(json.item.created_at),
-          updated_at: new Date(json.item.updated_at)
-        }
+          updated_at: new Date(json.item.updated_at),
+        },
       }
     }
     return json
@@ -97,22 +104,34 @@ const create = async (projectUuid: string, clipInput: ClipInput) => {
 }
 
 export default {
-  all: async (projectUuid: string, page: number, pageSize: number): Promise<PaginationResponseV2<Clip> | ErrorResponseV2> => {
+  all: async (
+    projectUuid: string,
+    page: number,
+    pageSize: number,
+  ): Promise<PaginationResponseV2<Clip> | ErrorResponseV2> => {
     try {
-      const response = await UtilV2.get(`projects/${projectUuid}/clips?page=${page}${pageSize ? `&page_size=${pageSize}` : ''}`)
+      const response = await UtilV2.get(
+        `projects/${projectUuid}/clips?page=${page}${
+          pageSize ? `&page_size=${pageSize}` : ''
+        }`,
+      )
       const json = await response.json()
-      if (json.success) json.items.map(item => ({
-        ...item,
-        created_at: new Date(item.created_at),
-        updated_at: new Date(item.updated_at)
-      }))
+      if (json.success)
+        json.items.map((item) => ({
+          ...item,
+          created_at: new Date(item.created_at),
+          updated_at: new Date(item.updated_at),
+        }))
       return json
     } catch (e) {
       return UtilV2.errorResponse(e)
     }
   },
-  
-  get: async (projectUuid: string, uuid: string): Promise<ReadResponseV2<Clip> | ErrorResponseV2> => {
+
+  get: async (
+    projectUuid: string,
+    uuid: string,
+  ): Promise<ReadResponseV2<Clip> | ErrorResponseV2> => {
     try {
       const response = await UtilV2.get(`projects/${projectUuid}/clips/${uuid}`)
       let json = await response.json()
@@ -122,8 +141,8 @@ export default {
           item: {
             ...json.item,
             created_at: new Date(json.item.created_at),
-            updated_at: new Date(json.item.updated_at)
-          }
+            updated_at: new Date(json.item.updated_at),
+          },
         }
       }
       return json
@@ -131,18 +150,24 @@ export default {
       return UtilV2.errorResponse(e)
     }
   },
-  
 
-  
-  createAsync: async (projectUuid: string, clipInput: AsyncClipInput): Promise<WriteResponseV2<Clip> | ErrorResponseV2> => {
+  createAsync: async (
+    projectUuid: string,
+    clipInput: AsyncClipInput,
+  ): Promise<WriteResponseV2<Clip> | ErrorResponseV2> => {
     return create(projectUuid, clipInput)
   },
 
-  createSync: async (projectUuid: string, clipInput: SyncClipInput): Promise<WriteResponseV2<Clip> | ErrorResponseV2> => {
+  createSync: async (
+    projectUuid: string,
+    clipInput: SyncClipInput,
+  ): Promise<WriteResponseV2<Clip> | ErrorResponseV2> => {
     return create(projectUuid, clipInput)
   },
 
-  createDirect: async (clipInput: DirectClipInput): Promise<DirectClip | DirectClipError | ErrorResponseV2> => {
+  createDirect: async (
+    clipInput: DirectClipInput,
+  ): Promise<DirectClip | DirectClipError | ErrorResponseV2> => {
     try {
       const response = await UtilV2.post('synthesize', clipInput, true)
       let json = await response.json()
@@ -151,50 +176,77 @@ export default {
       return UtilV2.errorResponse(e)
     }
   },
-  
-  // stream: async function* (streamInput, bufferSize = DEFAULT_BUFFER_SIZE, ignoreWavHeader = true): AsyncGenerator {
-  //   try {
-  //     const response = await UtilV2.post('stream', streamInput, true)
-  
-  //     // check for error response
-  //     if (!response.ok) {
-  //       const isJson = response.headers.get('content-type')?.includes('application/json');
-  //       const data = isJson ? await response.json() : null;
-  //       const error = (data && data.message) || response.status;
-  //       throw Error(error);
-  //     }
-  
-  //     const streamDecoder = new StreamDecoder(bufferSize, ignoreWavHeader)
-  //     streamDecoder.reset()
-  
-  //     // Iterate over the stream and start decoding, and returning data
-  //     for await (const chunk of response.body!) {
-  //       streamDecoder.decodeChunk(chunk)
-  //       const buffer = streamDecoder.flushBuffer()
-  //       if (buffer !== null)
-  //         yield buffer
-  //     }
-  
-  //     // Keep draining the buffer until the buffer.length < bufferSize or buffer.length == 0
-  //     let buffer = streamDecoder.flushBuffer()
-  //     while (buffer !== null) {
-  //       const buffToReturn = Buffer.from(buffer)
-  //       buffer = streamDecoder.flushBuffer()
-  //       yield buffToReturn
-  //     }
-  
-  //     // Drain any leftover content in the buffer, buffer.length will always be less than bufferSize here
-  //     buffer = streamDecoder.flushBuffer(true)
-  //     if (buffer !== null)
-  //       yield buffer
-  //   } catch (e) {
-  //     return UtilV2.errorResponse(e)
-  //   }
-  // }
-  
-  updateAsync: async (projectUuid: string, uuid: string, clipInput: AsyncClipInput) => {
+
+  stream: async function* (
+    streamInput: {
+      data: string
+      project_uuid: string
+      voice_uuid: string
+    },
+    bufferSize = DEFAULT_BUFFER_SIZE,
+    ignoreWavHeader = true,
+  ): AsyncGenerator {
     try {
-      const response = await UtilV2.put(`projects/${projectUuid}/clips/${uuid}`, clipInput)
+      const response = await UtilV2.post('stream', streamInput, true)
+
+      // check for error response
+      if (!response.ok || !response.body) {
+        const isJson = response.headers
+          .get('content-type')
+          ?.includes('application/json')
+        const data = isJson ? await response.json() : null
+        const error = (data && data.message) || response.status
+        throw Error(error)
+      }
+
+      const streamDecoder = new StreamDecoder(bufferSize, ignoreWavHeader)
+      streamDecoder.reset()
+
+      const reader = response.body.getReader()
+
+      // Iterate over the stream and start decoding, and returning data
+
+      try {
+        while (true) {
+          const { done, value } = await reader.read()
+          if (done) break
+
+          streamDecoder.decodeChunk(value)
+          const buffer = streamDecoder.flushBuffer()
+          if (buffer !== null) yield buffer
+        }
+      } finally {
+        reader.releaseLock()
+      }
+
+      // Keep draining the buffer until the buffer.length < bufferSize or buffer.length == 0
+      let buffer = streamDecoder.flushBuffer()
+      while (buffer !== null) {
+        const buffToReturn = Buffer.from(buffer)
+        buffer = streamDecoder.flushBuffer()
+        yield buffToReturn
+      }
+
+      // Drain any leftover content in the buffer, buffer.length will always be less than bufferSize here
+      buffer = streamDecoder.flushBuffer(true)
+      if (buffer !== null) yield buffer
+    } catch (e) {
+      // If an error occurs and the catch block is executed, the function will return a plain object (UtilV2.errorResponse(e)).
+      // This will cause the function to not return an async iterable, leading to an error, so we need to throw the error
+      throw e
+    }
+  },
+
+  updateAsync: async (
+    projectUuid: string,
+    uuid: string,
+    clipInput: AsyncClipInput,
+  ) => {
+    try {
+      const response = await UtilV2.put(
+        `projects/${projectUuid}/clips/${uuid}`,
+        clipInput,
+      )
       let json = await response.json()
       if (json.success) {
         json = {
@@ -202,8 +254,8 @@ export default {
           item: {
             ...json.item,
             created_at: new Date(json.item.created_at),
-            updated_at: new Date(json.item.updated_at)
-          }
+            updated_at: new Date(json.item.updated_at),
+          },
         }
       }
       return json
@@ -211,14 +263,16 @@ export default {
       return UtilV2.errorResponse(e)
     }
   },
-  
+
   destroy: async (projectUuid: string, uuid: string) => {
     try {
-      const response = await UtilV2.delete(`projects/${projectUuid}/clips/${uuid}`)
+      const response = await UtilV2.delete(
+        `projects/${projectUuid}/clips/${uuid}`,
+      )
       const json = response.json()
       return json
     } catch (e) {
       return UtilV2.errorResponse(e)
     }
-  }
+  },
 }