+++ src/v2/projects.ts
@@ -1,4 +1,11 @@
-import UtilV2, { DeleteResponseV2, ErrorResponseV2, PaginationResponseV2, ReadResponseV2, UpdateResponseV2, WriteResponseV2 } from "./util"
+import UtilV2, {
+  DeleteResponseV2,
+  ErrorResponseV2,
+  PaginationResponseV2,
+  ReadResponseV2,
+  UpdateResponseV2,
+  WriteResponseV2,
+} from './util'
 
 export interface Project {
   uuid: string
@@ -20,22 +27,30 @@ export interface ProjectInput {
 }
 
 export default {
-  all: async (page: number, pageSize?: number): Promise<PaginationResponseV2<Project> | ErrorResponseV2> => {
+  all: async (
+    page: number,
+    pageSize?: number,
+  ): Promise<PaginationResponseV2<Project> | ErrorResponseV2> => {
     try {
-      const response = await UtilV2.get(`projects?page=${page}${pageSize ? `&page_size=${pageSize}` : ''}`)
+      const response = await UtilV2.get(
+        `projects?page=${page}${pageSize ? `&page_size=${pageSize}` : ''}`,
+      )
       const json = await response.json()
-      if (json.success) json.items.map(item => ({
-        ...item,
-        created_at: new Date(item.created_at),
-        updated_at: new Date(item.updated_at)
-      }))
+      if (json.success)
+        json.items.map((item) => ({
+          ...item,
+          created_at: new Date(item.created_at),
+          updated_at: new Date(item.updated_at),
+        }))
       return json
     } catch (e) {
       return UtilV2.errorResponse(e)
     }
   },
 
-  get: async (uuid: string): Promise<ReadResponseV2<Project> | ErrorResponseV2> => {
+  get: async (
+    uuid: string,
+  ): Promise<ReadResponseV2<Project> | ErrorResponseV2> => {
     try {
       const response = await UtilV2.get(`projects/${uuid}`)
       let json = await response.json()
@@ -45,8 +60,8 @@ export default {
           item: {
             ...json.item,
             created_at: new Date(json.item.created_at),
-            updated_at: new Date(json.item.updated_at)
-          }
+            updated_at: new Date(json.item.updated_at),
+          },
         }
       }
       return json
@@ -54,8 +69,10 @@ export default {
       return UtilV2.errorResponse(e)
     }
   },
-  
-  create: async (projectInput: ProjectInput): Promise<WriteResponseV2<Project> | ErrorResponseV2> => {
+
+  create: async (
+    projectInput: ProjectInput,
+  ): Promise<WriteResponseV2<Project> | ErrorResponseV2> => {
     try {
       const response = await UtilV2.post('projects', projectInput)
       let json = await response.json()
@@ -65,8 +82,8 @@ export default {
           item: {
             ...json.item,
             created_at: new Date(json.item.created_at),
-            updated_at: new Date(json.item.updated_at)
-          }
+            updated_at: new Date(json.item.updated_at),
+          },
         }
       }
       return json
@@ -75,7 +92,10 @@ export default {
     }
   },
 
-  update: async (uuid: string, projectInput: ProjectInput): Promise<UpdateResponseV2<Project> | ErrorResponseV2> => {
+  update: async (
+    uuid: string,
+    projectInput: ProjectInput,
+  ): Promise<UpdateResponseV2<Project> | ErrorResponseV2> => {
     try {
       const response = await UtilV2.put(`projects/${uuid}`, projectInput)
       let json = await response.json()
@@ -85,8 +105,8 @@ export default {
           item: {
             ...json.item,
             created_at: new Date(json.item.created_at),
-            updated_at: new Date(json.item.updated_at)
-          }
+            updated_at: new Date(json.item.updated_at),
+          },
         }
       }
       return json
@@ -95,7 +115,9 @@ export default {
     }
   },
 
-  destroy: async (uuid: string): Promise<DeleteResponseV2 | ErrorResponseV2> => {
+  destroy: async (
+    uuid: string,
+  ): Promise<DeleteResponseV2 | ErrorResponseV2> => {
     try {
       const response = await UtilV2.delete(`projects/${uuid}`)
       const json = response.json()
@@ -103,6 +125,5 @@ export default {
     } catch (e) {
       return UtilV2.errorResponse(e)
     }
-  }
-
+  },
 }