+++ src/v2/recordings.ts
@@ -1,7 +1,14 @@
 import fetch from 'isomorphic-fetch'
 import FormData from 'isomorphic-form-data'
 import { context } from '../context'
-import UtilV2, { ErrorResponseV2, PaginationResponseV2, UpdateResponseV2, DeleteResponseV2, ReadResponseV2, WriteResponseV2 } from "./util"
+import UtilV2, {
+  ErrorResponseV2,
+  PaginationResponseV2,
+  UpdateResponseV2,
+  DeleteResponseV2,
+  ReadResponseV2,
+  WriteResponseV2,
+} from './util'
 
 export interface Recording {
   uuid: string
@@ -14,7 +21,6 @@ export interface Recording {
   updated_at: Date
 }
 
-
 export interface RecordingInput {
   name: string
   text: string
@@ -23,24 +29,38 @@ export interface RecordingInput {
 }
 
 export default {
-  all: async (voiceUuid: string, page: number, pageSize: number): Promise<PaginationResponseV2<Recording> | ErrorResponseV2> => {
+  all: async (
+    voiceUuid: string,
+    page: number,
+    pageSize: number,
+  ): Promise<PaginationResponseV2<Recording> | ErrorResponseV2> => {
     try {
-      const response = await UtilV2.get(`voices/${voiceUuid}/recordings?page=${page}${pageSize ? `&page_size=${pageSize}` : ''}`)
+      const response = await UtilV2.get(
+        `voices/${voiceUuid}/recordings?page=${page}${
+          pageSize ? `&page_size=${pageSize}` : ''
+        }`,
+      )
       const json = await response.json()
-      if (json.success) json.items.map(item => ({
-        ...item,
-        created_at: new Date(item.created_at),
-        updated_at: new Date(item.updated_at)
-      }))
+      if (json.success)
+        json.items.map((item) => ({
+          ...item,
+          created_at: new Date(item.created_at),
+          updated_at: new Date(item.updated_at),
+        }))
       return json
     } catch (e) {
       return UtilV2.errorResponse(e)
     }
   },
-  
-  get: async (voiceUuid: string, uuid: string): Promise<ReadResponseV2<Recording> | ErrorResponseV2> => {
+
+  get: async (
+    voiceUuid: string,
+    uuid: string,
+  ): Promise<ReadResponseV2<Recording> | ErrorResponseV2> => {
     try {
-      const response = await UtilV2.get(`voices/${voiceUuid}/recordings/${uuid}`)
+      const response = await UtilV2.get(
+        `voices/${voiceUuid}/recordings/${uuid}`,
+      )
       let json = await response.json()
       if (json.success) {
         json = {
@@ -48,8 +68,8 @@ export default {
           item: {
             ...json.item,
             created_at: new Date(json.item.created_at),
-            updated_at: new Date(json.item.updated_at)
-          }
+            updated_at: new Date(json.item.updated_at),
+          },
         }
       }
       return json
@@ -57,8 +77,13 @@ export default {
       return UtilV2.errorResponse(e)
     }
   },
-  
-  create: async (voiceUuid: string, recordingInput: RecordingInput, buffer: Buffer, fileSizeInBytes: number): Promise<WriteResponseV2<Recording> | ErrorResponseV2> => {
+
+  create: async (
+    voiceUuid: string,
+    recordingInput: RecordingInput,
+    buffer: Buffer,
+    fileSizeInBytes: number,
+  ): Promise<WriteResponseV2<Recording> | ErrorResponseV2> => {
     try {
       const formData = new FormData()
       formData.append('name', recordingInput.name)
@@ -66,17 +91,20 @@ export default {
       formData.append('emotion', recordingInput.emotion)
       formData.append('is_active', recordingInput.is_active ? 'true' : 'false')
       formData.append('file', buffer, { knownLength: fileSizeInBytes })
-  
-      const response = await fetch(context.endpoint('v2', `voices/${voiceUuid}/recordings`), {
-        method: 'POST',
-        headers: {
-          Authorization: context.headers().Authorization,
-          'Content-Type': 'multipart/form-data',
-          ...formData.getHeaders ? formData.getHeaders() : {}
+
+      const response = await fetch(
+        context.endpoint('v2', `voices/${voiceUuid}/recordings`),
+        {
+          method: 'POST',
+          headers: {
+            Authorization: context.headers().Authorization,
+            'Content-Type': 'multipart/form-data',
+            ...(formData.getHeaders ? formData.getHeaders() : {}),
+          },
+          body: formData,
         },
-        body: formData
-      })
-  
+      )
+
       let json = await response.json()
       if (json.success) {
         json = {
@@ -84,8 +112,8 @@ export default {
           item: {
             ...json.item,
             created_at: new Date(json.item.created_at),
-            updated_at: new Date(json.item.updated_at)
-          }
+            updated_at: new Date(json.item.updated_at),
+          },
         }
       }
       return json
@@ -93,10 +121,17 @@ export default {
       return UtilV2.errorResponse(e)
     }
   },
-  
-  update: async (voiceUuid: string, uuid: string, recordingInput: RecordingInput): Promise<UpdateResponseV2<Recording> | ErrorResponseV2> => {
+
+  update: async (
+    voiceUuid: string,
+    uuid: string,
+    recordingInput: RecordingInput,
+  ): Promise<UpdateResponseV2<Recording> | ErrorResponseV2> => {
     try {
-      const response = await UtilV2.put(`voices/${voiceUuid}/recordings/${uuid}`, recordingInput)
+      const response = await UtilV2.put(
+        `voices/${voiceUuid}/recordings/${uuid}`,
+        recordingInput,
+      )
       let json = await response.json()
       if (json.success) {
         json = {
@@ -104,8 +139,8 @@ export default {
           item: {
             ...json.item,
             created_at: new Date(json.item.created_at),
-            updated_at: new Date(json.item.updated_at)
-          }
+            updated_at: new Date(json.item.updated_at),
+          },
         }
       }
       return json
@@ -113,16 +148,19 @@ export default {
       return UtilV2.errorResponse(e)
     }
   },
-  
-  destroy: async (voiceUuid: string, uuid: string): Promise<DeleteResponseV2 | ErrorResponseV2> => {
+
+  destroy: async (
+    voiceUuid: string,
+    uuid: string,
+  ): Promise<DeleteResponseV2 | ErrorResponseV2> => {
     try {
-      const response = await UtilV2.delete(`voices/${voiceUuid}/recordings/${uuid}`)
+      const response = await UtilV2.delete(
+        `voices/${voiceUuid}/recordings/${uuid}`,
+      )
       const json = response.json()
       return json
     } catch (e) {
       return UtilV2.errorResponse(e)
     }
-  }
-
+  },
 }
-