+++ CHANGELOG.md
@@ -5,5 +5,5 @@ All notable changes to this project will be documented in this file.
 ## v2.0.3 - 05-23-2023
 
 ### Added
-- Updated the Voice creation function to accept a `consent` audio file in accordance with the requirements for the API
 
+- Updated the Voice creation function to accept a `consent` audio file in accordance with the requirements for the API